🎸 Bard: [documentation update] Fix Quick Start documentation examples (#3158)
- Replaced references to missing `genesis.chs` with `frankenstein_lab.chs` in `chimera-lang/README.md`.
- Removed incorrect `cd chimera-lang` from root `README.md` to prevent pathing confusion.
- Fixed `mad_scientist.prl` script to use a safe `1 push` and `drop` sequence instead of a crashing syntax, preventing an immediate crash when users run the Quick Start example.
- Fixed unrelated clippy warnings in `quipu-serializer`, `stego-attack`, and `system-attractor`.
- Logged these corrections in `.jules/bard.md`.

+++ experiments/stego-attack/src/stego.rs
@@ -75,7 +75,7 @@ pub fn extract(image: &RgbaImage) -> Result<Vec<u8>, String> {
     }
 
     // Sanity check length
-    let capacity_bytes = (image.width() * image.height() * 4 / 8) as u32 - 4;
+    let capacity_bytes = (image.width() * image.height() * 4 / 8) - 4;
     if len > capacity_bytes {
         return Err(format!(
             "Extracted length {} exceeds capacity {}",
